+++ opentelemetry-sdk/src/opentelemetry/sdk/_logs/_internal/export/__init__.py
@@ -20,6 +20,7 @@ import logging
 import os
 import sys
 import threading
+import weakref
 from os import environ, linesep
 from typing import IO, Callable, Deque, Optional, Sequence
 
@@ -214,7 +215,8 @@ class BatchLogRecordProcessor(LogRecordProcessor):
         self._worker_awaken = threading.Event()
         self._worker_thread.start()
         if hasattr(os, "register_at_fork"):
-            os.register_at_fork(after_in_child=self._at_fork_reinit)  # pylint: disable=protected-access
+            weak_reinit = weakref.WeakMethod(self._at_fork_reinit)
+            os.register_at_fork(after_in_child=lambda: weak_reinit()())  # pylint: disable=unnecessary-lambda
         self._pid = os.getpid()
 
     def _should_export_batch(